Move playwright nav and item extraction logic to separate modules Improve scraping behaviour; add scope fallback to list_route.el => el.frame key if exists but list_items.frame doesn't
